nexus/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# GraphQL
strawberry-graphql[fastapi]==0.215.1

# HTTP Client
httpx==0.25.2

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Database & Caching
redis==5.0.1
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.23
alembic==1.13.1

# Retry Logic
tenacity==8.2.3

# Logging
structlog==23.2.0
python-json-logger==2.0.7

# Monitoring
prometheus-client==0.19.0

# Utilities
python-multipart==0.0.6
python-dotenv==1.0.0
typing-extensions==4.8.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
